# Get index of first non ascii byte