--- a/src/libraries/System.Private.CoreLib/src/System/Text/ASCIIUtility.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Text/ASCIIUtility.cs
@@ -476,7 +476,7 @@ namespace System.Text
                 if (!ContainsNonAsciiByte_AdvSimd(currentAdvSimdIndex))
                 {
                     pBuffer += SizeOfVector128;
-                    currentAdvSimdIndex = secondSseMask;
+                    currentAdvSimdIndex = secondAdvSimdIndex;
                 }
             }
             else