Can be reviewed now. I'll add perf numbers in a bit.

Implements `GetIndexOfFirstNonAsciiByte` from https://github.com/dotnet/runtime/issues/35034