@pgovind 2020-07-30
Perf Results:

```
summary:
better: 2, geomean: 1.922
total diff: 2

No Slower results for the provided threshold = 2% and noise filter = 0.3ns.

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Text.Experimental.Perf.IsAscii |      1.41 |          1838.15 |          1302.61 | |

No file given
```