@kunalspathak 2020-08-04
Could you please rerun the benchmark and share the latest numbers?