@kunalspathak 2020-08-10
Do you mind sending a PR to `dotnet/performance` to add `IsAscii()` benchmark? I don't see it there. Also, can you please share the code of benchmark?