@pgovind 2020-08-10
> Do you mind sending a PR to dotnet/performance to add IsAscii() benchmark? I don't see it there. Also, can you please share the code of benchmark?

Here's the local benchmark I was using https://github.com/dotnet/performance/pull/1445