@pgovind 2020-08-13
Latest perf numbers on my machine running WSL2 (I couldn't get it to run on Windows, so this is the next best thing):

```
No Slower results for the provided threshold = 0.1% and noise filter = 0.3ns.

| Faster                                                            | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ----------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Text.Experimental.Perf.ToUtf16_nonascii_110                |      3.90 |        664380.29 |        170197.37 |         |
| System.Text.Experimental.Perf.IsAscii_GetIndexOfFirstNonAsciiByte |      3.64 |        134643.28 |         36941.30 |         |
| System.Text.Experimental.Perf.IsNormalized_WidenAsciiToUtf16      |      2.72 |       1280532.46 |        470127.60 | bimodal |
| System.Text.Experimental.Perf.ToUtf16_nonascii_cyrillic           |      2.51 |        393356.80 |        156837.70 |         |
| System.Text.Experimental.Perf.ToUtf16_nonascii_chinese            |      1.87 |       1222065.89 |        654866.60 |         |

No file given
```

These numbers high enough to leave me a little bit skeptical. The `GetIndexOfFirstNonAsciiByte` test cases is a long string with no non-ascii bytes though, so it's not inconceivable that we're seeing significant gains(The WidenAscii case has a bi-modal distribution, so take those numbers with a pinch of salt). I'm not sure how being on WSL2 affects perf. I am however reasonably convinced that we are seeing some perf gains from this PR. Once we branch off on Friday and produce an official build, I can re-run this benchmark and get numbers. Because the  methods being measured here call each other and affect perf, those numbers will be a much better representation of the overall perf gain.

cc @kunalspathak