@kunalspathak 2020-08-13
Thanks @pgovind . I went through your benchmark PR https://github.com/dotnet/performance/pull/1445 and added some comments to eliminate any other noise. I am OK merging it at this point. Just want to double check with @echesakovMSFT  on his thoughts on running benchmarks on WSL2.