@kunalspathak 2020-08-13
If you don't mind, can you have open a follow-up issue to optimize the code as per Tamar's suggestions?