@pgovind 2020-08-13
Yup, filed https://github.com/dotnet/runtime/issues/40805